@@ -1,93 +1,40 @@
 {
   "name": "kmcmgmt",
-  "version": "0.1.0",
   "private": true,
-  "homepage": ".",
-  "dependencies": {
-    "@babel/traverse": "7.23.2",
-    "@emotion/react": "^11.9.3",
-    "@emotion/styled": "^11.9.3",
-    "@fontsource/roboto": "^4.5.7",
-    "@fontsource/roboto-mono": "^4.5.7",
-    "@mui/icons-material": "^5.8.4",
-    "@mui/material": "^5.8.4",
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^13.3.0",
-    "@testing-library/user-event": "^13.5.0",
-    "axios": "^0.28.0",
-    "babel-loader": "^9.1.0",
-    "buffer": "^6.0.3",
-    "formik": "^2.2.9",
-    "json5": "^2.2.2",
-    "loader-utils": "^2.0.4",
-    "material-components-web": "^14.0.0",
-    "notistack": "^2.0.5",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-scripts": "^5.0.1",
-    "react-table": "^7.8.0",
-    "react-test-renderer": "^18.2.0",
-    "web-vitals": "^2.1.4",
-    "yup": "^0.32.11"
-  },
+  "version": "0.1.0",
+  "type": "module",
   "scripts": {
-    "start": "PORT=3001 react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "dev": "vite",
+    "build": "vite build",
+    "lint": "eslint .",
+    "preview": "vite preview",
+    "test": "vitest"
   },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ],
-    "overrides": [
-      {
-        "files": [
-          "**/*.stories.*"
-        ],
-        "rules": {
-          "import/no-anonymous-default-export": "off"
-        }
-      }
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+  "dependencies": {
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.0",
+    "@fontsource/roboto": "^5.2.5",
+    "@fontsource/roboto-mono": "^5.2.5",
+    "@mui/icons-material": "^6.4.8",
+    "@mui/material": "^6.4.8",
+    "axios": "^1.8.4",
+    "formik": "^2.4.6",
+    "notistack": "^3.0.2",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "react-table": "^7.8.0",
+    "yup": "^1.6.1"
   },
   "devDependencies": {
-    "@babel/plugin-transform-private-property-in-object": "^7.25.9",
-    "babel-plugin-named-exports-order": "^0.0.2",
-    "http-proxy-middleware": "^2.0.7",
-    "prop-types": "^15.8.1",
-    "webpack": "^5.76.0"
-  },
-  "resolutions": {
-    "loader-utils": "2.0.4",
-    "minimatch": "^3.0.5",
-    "json5": "^2.2.2",
-    "debug": "^3.1.0",
-    "css-select": "^5.1.0",
-    "ejs": "3.1.10",
-    "@babel/core": "^7.23.2",
-    "@babel/traverse": "^7.23.2",
-    "webpack-dev-middleware": "^5.3.4",
-    "cross-spawn": "^7.0.5",
-    "http-proxy-middleware": "^2.0.7",
-    "rollup": "^2.79.2",
-    "body-parser": "^1.20.3",
-    "path-to-regexp": "^0.1.12",
-    "ws": "^8.17.1",
-    "braces": "^3.0.3",
-    "postcss": "^8.4.31"
+    "@eslint/js": "^9.21.0",
+    "@types/react": "^19.0.10",
+    "@types/react-dom": "^19.0.4",
+    "@vitejs/plugin-react": "^4.3.4",
+    "eslint": "^9.21.0",
+    "eslint-plugin-react-hooks": "^5.1.0",
+    "eslint-plugin-react-refresh": "^0.4.19",
+    "globals": "^15.15.0",
+    "vite": "^6.2.0",
+    "vitest": "^3.0.9"
   }
 }
